Refactor statefulset pkg list function
